FIDO-Dynamic-Components/Fuzzware/pipeline/requirements.txt
click==7.1.2
rq
PyYAML>=5.3
watchdog
intelhex
pyelftools
